Temporary ugly version just to check it wont break on VPS

+++ frontend/app/courses/promo/[slug]/page.tsx
@@ -454,6 +454,7 @@ export default function CoursePromoPage() {
         "Вы действительно хотите удалить свой отзыв? Это действие нельзя отменить.",
       onConfirm: async () => {
         try {
+          if (!reviewStats.user_review) return;
           await apiFetch(`/reviews/${reviewStats.user_review.id}`, {
             method: "DELETE",
             headers: { Authorization: `Bearer ${token}` },
